# MQE: add support for `count_values`