--- a/pkg/streamingpromql/benchmarks/benchmarks.go
+++ b/pkg/streamingpromql/benchmarks/benchmarks.go
@@ -258,10 +258,10 @@ func TestCases(metricSizes []int) []BenchCase {
 		{
 			Expr: "avg by (l)(nh_X)",
 		},
-		//{
-		//	Expr: "count_values('value', h_X)",
-		//  Steps: 100,
-		//},
+		{
+			Expr:  "count_values('value', h_X)",
+			Steps: 100,
+		},
 		{
 			Expr: "topk(1, a_X)",
 		},