--- a/pkg/streamingpromql/operators/aggregations/count_values.go
+++ b/pkg/streamingpromql/operators/aggregations/count_values.go
@@ -1,0 +1,227 @@
+// SPDX-License-Identifier: AGPL-3.0-only
+// Provenance-includes-location: https://github.com/prometheus/prometheus/blob/main/promql/engine.go
+// Provenance-includes-license: Apache-2.0
+// Provenance-includes-copyright: The Prometheus Authors
+
+package aggregations
+
+import (
+	"context"
+	"fmt"
+	"slices"
+	"strconv"
+
+	"github.com/prometheus/common/model"
+	"github.com/prometheus/prometheus/model/labels"
+	"github.com/prometheus/prometheus/promql"
+	"github.com/prometheus/prometheus/promql/parser/posrange"
+
+	"github.com/grafana/mimir/pkg/streamingpromql/limiting"
+	"github.com/grafana/mimir/pkg/streamingpromql/types"
+)
+
+type CountValues struct {
+	Inner                    types.InstantVectorOperator
+	LabelName                types.StringOperator
+	TimeRange                types.QueryTimeRange
+	Grouping                 []string // If this is a 'without' aggregation, NewAggregation will ensure that this slice contains __name__.
+	Without                  bool
+	MemoryConsumptionTracker *limiting.MemoryConsumptionTracker
+
+	expressionPosition posrange.PositionRange
+
+	resolvedLabelName string
+
+	series [][]promql.FPoint
+
+	// Reuse instances used to generate series labels rather than recreating them every time.
+	labelsBuilder     *labels.Builder
+	labelsBytesBuffer []byte
+}
+
+var _ types.InstantVectorOperator = &CountValues{}
+
+func NewCountValues(
+	inner types.InstantVectorOperator,
+	labelName types.StringOperator,
+	timeRange types.QueryTimeRange,
+	grouping []string,
+	without bool,
+	memoryConsumptionTracker *limiting.MemoryConsumptionTracker,
+	expressionPosition posrange.PositionRange,
+) *CountValues {
+	return &CountValues{
+		Inner:                    inner,
+		LabelName:                labelName,
+		TimeRange:                timeRange,
+		Grouping:                 grouping,
+		Without:                  without,
+		MemoryConsumptionTracker: memoryConsumptionTracker,
+		expressionPosition:       expressionPosition,
+	}
+}
+
+type countValuesSeries struct {
+	labels           labels.Labels
+	outputPointCount int
+	count            []int // One entry per timestamp.
+}
+
+func (c *CountValues) SeriesMetadata(ctx context.Context) ([]types.SeriesMetadata, error) {
+	if err := c.loadLabelName(); err != nil {
+		return nil, err
+	}
+
+	innerMetadata, err := c.Inner.SeriesMetadata(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	defer types.PutSeriesMetadataSlice(innerMetadata)
+
+	c.labelsBuilder = labels.NewBuilder(labels.EmptyLabels())
+	c.labelsBytesBuffer = make([]byte, 0, 1024) // Why 1024 bytes? It's what labels.Labels.String() uses as a buffer size, so we use that as a sensible starting point too.
+	defer func() {
+		// Don't hold onto the instances used to generate series labels for longer than necessary.
+		c.labelsBuilder = nil
+		c.labelsBytesBuffer = nil
+	}()
+
+	accumulator := map[string]*countValuesSeries{}
+
+	for _, s := range innerMetadata {
+		data, err := c.Inner.NextSeries(ctx)
+		if err != nil {
+			return nil, err
+		}
+
+		for _, p := range data.Floats {
+			if err := c.incrementCount(s.Labels, p.T, strconv.FormatFloat(p.F, 'f', -1, 64), accumulator); err != nil {
+				return nil, err
+			}
+		}
+
+		for _, p := range data.Histograms {
+			if err := c.incrementCount(s.Labels, p.T, p.H.String(), accumulator); err != nil {
+				return nil, err
+			}
+		}
+
+		types.PutInstantVectorSeriesData(data, c.MemoryConsumptionTracker)
+	}
+
+	outputMetadata := types.GetSeriesMetadataSlice(len(accumulator))
+	c.series = make([][]promql.FPoint, 0, len(accumulator))
+
+	for _, s := range accumulator {
+		outputMetadata = append(outputMetadata, types.SeriesMetadata{Labels: s.labels})
+
+		points, err := s.ToPoints(c.MemoryConsumptionTracker, c.TimeRange)
+		if err != nil {
+			return nil, err
+		}
+
+		c.series = append(c.series, points)
+
+		types.IntSlicePool.Put(s.count, c.MemoryConsumptionTracker)
+	}
+
+	return outputMetadata, nil
+}
+
+func (c *CountValues) loadLabelName() error {
+	c.resolvedLabelName = c.LabelName.GetValue()
+	if !model.LabelName(c.resolvedLabelName).IsValid() {
+		return fmt.Errorf("invalid label name %q for count_values", c.resolvedLabelName)
+	}
+
+	if c.Without {
+		c.Grouping = append(c.Grouping, labels.MetricName)
+	}
+
+	slices.Sort(c.Grouping)
+	return nil
+}
+
+func (c *CountValues) incrementCount(seriesLabels labels.Labels, t int64, value string, accumulator map[string]*countValuesSeries) error {
+	l := c.computeOutputLabels(seriesLabels, value)
+	c.labelsBytesBuffer = l.Bytes(c.labelsBytesBuffer)
+	series, exists := accumulator[string(c.labelsBytesBuffer)] // Important: don't extract the string(...) call here - passing it directly allows us to avoid allocating it.
+
+	if !exists {
+		series = &countValuesSeries{
+			labels: l,
+		}
+
+		var err error
+		series.count, err = types.IntSlicePool.Get(c.TimeRange.StepCount, c.MemoryConsumptionTracker)
+		if err != nil {
+			return err
+		}
+
+		series.count = series.count[:c.TimeRange.StepCount]
+		accumulator[string(c.labelsBytesBuffer)] = series
+	}
+
+	idx := c.TimeRange.PointIndex(t)
+
+	if series.count[idx] == 0 {
+		series.outputPointCount++
+	}
+
+	series.count[idx]++
+
+	return nil
+}
+
+func (c *CountValues) computeOutputLabels(seriesLabels labels.Labels, value string) labels.Labels {
+	c.labelsBuilder.Reset(seriesLabels)
+
+	if c.Without {
+		c.labelsBuilder.Del(c.Grouping...)
+	} else {
+		c.labelsBuilder.Keep(c.Grouping...)
+	}
+
+	c.labelsBuilder.Set(c.resolvedLabelName, value)
+
+	return c.labelsBuilder.Labels()
+}
+
+func (s *countValuesSeries) ToPoints(memoryConsumptionTracker *limiting.MemoryConsumptionTracker, timeRange types.QueryTimeRange) ([]promql.FPoint, error) {
+	p, err := types.FPointSlicePool.Get(s.outputPointCount, memoryConsumptionTracker)
+	if err != nil {
+		return nil, err
+	}
+
+	for idx, count := range s.count {
+		if count == 0 {
+			continue
+		}
+
+		t := timeRange.IndexTime(int64(idx))
+		p = append(p, promql.FPoint{T: t, F: float64(count)})
+	}
+
+	return p, nil
+}
+
+func (c *CountValues) NextSeries(_ context.Context) (types.InstantVectorSeriesData, error) {
+	if len(c.series) == 0 {
+		return types.InstantVectorSeriesData{}, types.EOS
+	}
+
+	p := c.series[0]
+	c.series = c.series[1:]
+
+	return types.InstantVectorSeriesData{Floats: p}, nil
+}
+
+func (c *CountValues) ExpressionPosition() posrange.PositionRange {
+	return c.expressionPosition
+}
+
+func (c *CountValues) Close() {
+	c.Inner.Close()
+	c.LabelName.Close()
+}