--- a/pkg/streamingpromql/testdata/ours/aggregators.test
+++ b/pkg/streamingpromql/testdata/ours/aggregators.test
@@ -578,3 +578,50 @@ eval range from 0 to 12m step 6m bottomk(2, series)
 
 eval range from 0 to 12m step 6m bottomk(1, series)
   series{case="-Inf"}  -Inf -Inf
+
+clear
+
+# Test count_values.
+load 6m
+  series{idx="0", case="Inf"}   Inf  Inf  Inf
+  series{idx="0", case="-Inf"}  -Inf -Inf _
+  series{idx="0", case="NaN"}   NaN  NaN  NaN
+  series{idx="0", case="1"}     1    1    _
+  series{idx="1", case="Inf"}   Inf  _    _
+  series{idx="1", case="-Inf"}  -Inf _    _
+  series{idx="1", case="NaN"}   NaN  _    NaN
+  series{idx="1", case="1"}     1    _    _
+
+eval range from 0 to 12m step 6m count_values("value", series)
+  {value="+Inf"} 2 1 1
+  {value="-Inf"} 2 1 _
+  {value="NaN"}  2 1 2
+  {value="1"}    2 1 _
+
+eval range from 0 to 12m step 6m count_values by (case) ("value", series)
+  {case="Inf", value="+Inf"}  2 1 1
+  {case="-Inf", value="-Inf"} 2 1 _
+  {case="NaN", value="NaN"}   2 1 2
+  {case="1", value="1"}       2 1 _
+
+eval range from 0 to 12m step 6m count_values by (idx) ("value", series)
+  {idx="0", value="+Inf"} 1 1 1
+  {idx="0", value="-Inf"} 1 1 _
+  {idx="0", value="NaN"}  1 1 1
+  {idx="0", value="1"}    1 1 _
+  {idx="1", value="+Inf"} 1 _ _
+  {idx="1", value="-Inf"} 1 _ _
+  {idx="1", value="NaN"}  1 _ 1
+  {idx="1", value="1"}    1 _ _
+
+# Overwrite grouping label with value.
+eval range from 0 to 12m step 6m count_values by (idx) ("idx", series)
+  {idx="+Inf"} 2 1 1
+  {idx="-Inf"} 2 1 _
+  {idx="NaN"}  2 1 2
+  {idx="1"}    2 1 _
+
+# Prometheus' engine quotes the label name twice: see https://github.com/prometheus/prometheus/pull/16054.
+# Once that fix is vendored into Mimir, we can remove the two (\\")? groups below.
+eval_fail instant at 0 count_values("a\xc5z", series)
+  expected_fail_regexp invalid label name "(\\")?a\\(\\)?xc5z(\\")?"( for count_values)?