… pkg/streamingpromql/operators/aggregations/count_values_test.go